Hosting WordPress Using Amazon S3
This lab demonstrates how to use Amazon S3 for static website hosting. In this lab, you convert a WordPress website that is running on an Amazon EC2 instance into a static website hosted on Amazon S3. WordPress is an open-source content management system that can be used to host web pages, blog posts, and even shopping websites. It runs more than 60 million websites on the Internet and is highly extensible via a plugin and templating system.